Syntax highlighting fixes for quoting and functions
Quoting wouldn't handle backslashed quotes (\") for nesting.  Added
that, and the rarer case of \\" which I found in a function call where I
was attempting to get a literal \" provided to a template.

Also added a rough hack to highlight functions such as split() or join()
differently than variables inside an interpolation (or outside).

Added the comment highlighting for the // comment type.

Added highlighting for tfvars files, using the basic terraform filetype.
This is sufficient to highlight strings and comments, which comprise the
whole tfvars file.