Make sure the command is called from the correct path
Terraform always tries to use relative paths[1][2] which can cause unexpected behavior in combination with how vim is automatically dereferencing symlinks.

You can see the problem by running `:pwd` and `:echo system('pwd')`. Both commands will show different paths which is the root cause of this problem.

This PR fixes the problem by making sure the command is always executed from the same directy as vim uses to reference the file.

Additionally I added `-no-color` to the command to prevent all the ANSI escape codes in any error output. The `-no-color` option has always been available for the `fmt` subcommand, but it just wasn't documented. I opened a PR for that as well on the Terraform repo (#)

[1] https://github.com/hashicorp/terraform/blob/master/command/fmt.go#L125
[2] https://github.com/hashicorp/terraform/blob/master/command/meta_config.go#L25