add support for hcl file extensions
Terragrunt uses the .hcl extension for configuring its tool, which
follows the HCL syntax. There is a terragrunt hclfmt command, but
it doesn't properly support stdin/stdout yet (though I have an
open issue for this on that repo).

In the short term, the terraform fmt command should be sufficient,
so this PR adds support for binding hcl files for this.