fix: set expandtab so that newlines respect indent
I've for a long time been dealing with issues when using heredocs and
tabs.  Today I notice that it too inserted tabs when creating newlines
inside a resource.  Have never really been bothers by it as these got
changed to spaces after running `terraform fmt`.  However, as said, I
noticed that this occurred more places and needed to look for a fix.

By explicitly setting `expandtab`, we make sure that Vim uses spaces
instead of tabs when we press enter to insert a newline.